added heuristic function evaluation
This evaluation method first extracts the positions of the two queens and the arrow from the game board. It then calculates the distance between the player's queen and the opponent's queen, the number of moves available to the opponent queen, the number of moves available to the player's queen, and the number of moves available to the arrow. Finally, it uses these values to compute a heuristic value for the current state of the game board.